requirements.txt
pytest==7.1.3
allure-pytest==2.13.2
requests==2.31.0